refactor v2 eval ref start tests
