+++ crates/rulemorph/src/v2_eval/tests/start_eval.rs
@@ -1,0 +1,145 @@
+use super::*;
+use serde_json::json;
+
+#[test]
+fn test_eval_start_literal_string() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Literal(json!("hello")),
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!("hello")));
+}
+
+#[test]
+fn test_eval_start_literal_number() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Literal(json!(42)),
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!(42)));
+}
+
+#[test]
+fn test_eval_start_literal_bool() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Literal(json!(true)),
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!(true)));
+}
+
+#[test]
+fn test_eval_start_literal_null() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Literal(json!(null)),
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!(null)));
+}
+
+#[test]
+fn test_eval_start_literal_array() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Literal(json!([1, 2, 3])),
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!([1, 2, 3])));
+}
+
+#[test]
+fn test_eval_start_literal_object() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Literal(json!({"key": "value"})),
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!({"key": "value"})));
+}
+
+#[test]
+fn test_eval_start_ref() {
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::Ref(V2Ref::Input("name".to_string())),
+        &json!({"name": "Bob"}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!("Bob")));
+}
+
+#[test]
+fn test_eval_start_pipe_value() {
+    let ctx = V2EvalContext::new().with_pipe_value(EvalValue::Value(json!(42)));
+    let result = eval_v2_start(
+        &V2Start::PipeValue,
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Value(v)) if v == json!(42)));
+}
+
+#[test]
+fn test_eval_start_pipe_value_not_available() {
+    // When pipe value is not set, it returns Missing (not error)
+    // This allows ops like lookup_first that don't use pipe input to work
+    let ctx = V2EvalContext::new();
+    let result = eval_v2_start(
+        &V2Start::PipeValue,
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(result.is_ok());
+    assert_eq!(result.unwrap(), EvalValue::Missing);
+}
+
+#[test]
+fn test_eval_start_pipe_value_missing() {
+    let ctx = V2EvalContext::new().with_pipe_value(EvalValue::Missing);
+    let result = eval_v2_start(
+        &V2Start::PipeValue,
+        &json!({}),
+        None,
+        &json!({}),
+        "test",
+        &ctx,
+    );
+    assert!(matches!(result, Ok(EvalValue::Missing)));
+}
